% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_of_lists.R
\name{is_list_of_lists}
\alias{is_list_of_lists}
\title{Check if an object is a list of lists}
\usage{
is_list_of_lists(object)
}
\arguments{
\item{object}{any R object (anything)}
}
\value{
TRUE if object is a list of lists, FALSE if it is not. Any non-list elements will cause this function to return FALSE
}
\description{
Check if an object is a list of lists
}
\seealso{
Other nested_lists: 
\code{\link{list_of_lists_retrieve_second_level_property}()}
}
\concept{nested_lists}
